Rename PackageArgs to ResourcePackageArgs. Copy functionality from ResourcePackager into PackagerService in prepartion for incremental changes coming next
